requirements.txt
PyQt6
ultralytics
torch
torchvision
Pillow